requirements.txt
openai-whisper
fastapi
uvicorn
pydub
